% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_compilation.R
\name{lv_add_membership}
\alias{lv_add_membership}
\title{Add compilation membership to the named columns of one dataset}
\usage{
lv_add_membership(L, tsids, compilation, version)
}
\description{
Add compilation membership to the named columns of one dataset
}
\keyword{internal}
