% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{lv_changelog_append}
\alias{lv_changelog_append}
\title{Append an entry to a dataset's changelog}
\usage{
lv_changelog_append(L, entry)
}
\arguments{
\item{L}{A LiPD object.}

\item{entry}{From \code{\link[=lv_changelog_entry]{lv_changelog_entry()}}.}
}
\value{
\code{L}, with the entry appended.
}
\description{
Append an entry to a dataset's changelog
}
