% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_sheet.R
\name{lv_col_letter}
\alias{lv_col_letter}
\title{Spreadsheet column letter for a 1-based index}
\usage{
lv_col_letter(i)
}
\arguments{
\item{i}{Column index.}
}
\description{
Spreadsheet column letter for a 1-based index
}
