% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{lv_config_all}
\alias{lv_config_all}
\title{Configuration for every compilation}
\usage{
lv_config_all(...)
}
\arguments{
\item{...}{Passed to \code{\link[=lv_config]{lv_config()}}.}
}
\value{
A named list of \code{lv_config}.
}
\description{
Useful as a preflight check: it validates all of them at once.
}
