% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_compilation.R
\name{lv_drop_membership}
\alias{lv_drop_membership}
\title{Remove compilation membership from the named columns}
\usage{
lv_drop_membership(L, tsids, compilation)
}
\description{
Remove compilation membership from the named columns
}
\keyword{internal}
