% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{lv_gc}
\alias{lv_gc}
\title{Prune old trash generations}
\usage{
lv_gc(dir = lv_path("database"), keep = 10)
}
\arguments{
\item{dir}{Database directory.}

\item{keep}{Number of most recent runs to retain.}
}
\value{
Number of generations removed.
}
\description{
Prune old trash generations
}
