% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{lv_ingest_walk}
\alias{lv_ingest_walk}
\title{Walk a LiPD object's columns in the same order as \code{\link[=lv_ingest_scan]{lv_ingest_scan()}}}
\usage{
lv_ingest_walk(L)
}
\description{
Walk a LiPD object's columns in the same order as \code{\link[=lv_ingest_scan]{lv_ingest_scan()}}
}
\keyword{internal}
