% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{lv_lock_status}
\alias{lv_lock_status}
\alias{lv_lock_break}
\title{Inspect a lock}
\usage{
lv_lock_status(compilation)

lv_lock_break(compilation)
}
\arguments{
\item{compilation}{Compilation name.}
}
\value{
A list; \code{locked = FALSE} when free.
}
\description{
Inspect a lock
}
