% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{lv_references_path}
\alias{lv_references_path}
\title{Where the reference database lives}
\usage{
lv_references_path(store = qc_store())
}
\arguments{
\item{store}{A QC store.}
}
\value{
A path.
}
\description{
Where the reference database lives
}
