% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{lv_replay_versions}
\alias{lv_replay_versions}
\alias{lv_replay_root}
\title{Version directories that can be replayed}
\usage{
lv_replay_versions(compilation, root = lv_replay_root())

lv_replay_root()
}
\arguments{
\item{compilation}{Compilation name.}

\item{root}{Directory of published compilation websites.}
}
\value{
A tibble of \code{version} and \code{path}, oldest first.
}
\description{
Version directories that can be replayed
}
