% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{lv_scan_diff}
\alias{lv_scan_diff}
\title{Compare two scans}
\usage{
lv_scan_diff(old, new)
}
\arguments{
\item{old, new}{\code{lv_scan} objects.}
}
\value{
A tibble of added / removed / changed files.
}
\description{
Compare two scans
}
