% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{lv_version_parse}
\alias{lv_version_parse}
\alias{lv_version_string}
\title{Parse and render version strings}
\usage{
lv_version_parse(v)

lv_version_string(publication, dataset, metadata)
}
\arguments{
\item{v}{A version string like \code{"1_0_3"}.}

\item{publication, dataset, metadata}{Components.}
}
\value{
An integer vector of \code{publication}, \code{dataset}, \code{metadata}.
}
\description{
Parse and render version strings
}
