% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab_review.R
\name{lv_vocab_candidates}
\alias{lv_vocab_candidates}
\title{Closest vocabulary terms to a value}
\usage{
lv_vocab_candidates(value, key, vocab = lv_vocab(), n = 5)
}
\arguments{
\item{value}{A single value.}

\item{key}{Vocabulary key.}

\item{vocab}{From \code{\link[=lv_vocab]{lv_vocab()}}.}

\item{n}{How many to return.}
}
\value{
A character vector of candidate \code{lipdName}s.
}
\description{
Closest vocabulary terms to a value
}
