% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{lv_with_run}
\alias{lv_with_run}
\title{Run a block with an active run context}
\usage{
lv_with_run(run_id, code)
}
\arguments{
\item{run_id}{Run identifier.}

\item{code}{Code to evaluate.}
}
\description{
Run a block with an active run context
}
