% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_merge.R
\name{qc_merge_policy}
\alias{qc_merge_policy}
\title{Merge policy}
\usage{
qc_merge_policy(strict = TRUE, clear_sentinel = LV_CLEAR_SENTINEL)
}
\arguments{
\item{strict}{Abort on unresolved conflicts.}

\item{clear_sentinel}{Value a curator types to clear a cell explicitly.}
}
\description{
Merge policy
}
