% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_sheet.R
\name{sheet_read}
\alias{sheet_read}
\title{Read one tab}
\usage{
sheet_read(backend, id, tab)
}
\arguments{
\item{backend}{A backend.}

\item{id}{Sheet id.}

\item{tab}{Tab name.}
}
\description{
Read one tab
}
