% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_sheet.R
\name{sheet_write}
\alias{sheet_write}
\title{Write one tab}
\usage{
sheet_write(backend, id, tab, x)
}
\arguments{
\item{backend}{A backend.}

\item{id}{Sheet id.}

\item{tab}{Tab name.}

\item{x}{A data frame.}
}
\description{
Write one tab
}
